addition of the item.py file to handle the items class & also the view model for the tests to run.
