+++ .env.test
@@ -7,12 +7,7 @@ FLASK_DEBUG=true
 
 # Required to use Flask's session feature (for storing todos in a browser cookie in the first exercise): https://flask.palletsprojects.com/en/2.3.x/config/#SECRET_KEY
 SECRET_KEY=secret-key
-TRELLO_BOARD_ID='658ec5346800f22eb024ccec'
 
-TRELLO_API_KEY=test_api_key
-TRELLO_API_TOKEN=test_api_token
-
-
-MONGODB_NAME=testdb
-MONGODB_COLLECTION_NAME=testcollectionname
+MONGODB_NAME=todoappdb
+MONGODB_COLLECTION_NAME=todo_items
 MONGODB_CONNECTIONSTRING=mongodb://fakemongo.com
