+++ pyproject.toml
@@ -7,12 +7,12 @@ authors = []
 [tool.poetry.dependencies]
 python = "^3.8"
 Flask = "^2.2.5"
-python-dotenv = "^0.14.0"
 requests = "^2.31.0"
 pytest = "^8.0.0"
 gunicorn = "^21.2.0"
 pymongo = "^4.8.0"
 mongomock = "^4.1.2"
+python-dotenv = "^1.0.1"
 
 [tool.poetry.dev-dependencies]
